+++ frontend/src/components/index.js
@@ -1,0 +1,2 @@
+export { Filter } from "./Filter";
+export { Sort } from "./Sort";
