--- frontend/src/data.js
@@ -1,308 +1,0 @@
-import {v4 as uuidv4} from "uuid";
-const data=[
-/*----------------------brinjal-seed---------------- */
-  {
-    id: uuidv4(),
-    name: "",
-    image: "",
-    price: ,
-    material: "seed",
-    brand: "",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: ,
-    offer: 10,
-    idealFor:"brinjal" ,
-    color: "purple"
-  },
-  {
-    id: uuidv4(),
-    name: "LALITA BRINJAL EASTWEST F1 HYBRID 10GM",
-    image: "https://farmagritech.com/wp-content/uploads/2019/12/Photo-from-siddaraddi-ramaraddi-11.jpg",
-    price: 130.00,
-    material:"seed" ,
-    brand: "Lalita",
-    inStock: true,
-    isnew: true ,
-    fastDelivery: true,
-    ratings: 3,
-    offer: 6,
-    idealFor:"brinjal", 
-    color: "purple"
-  },
-
-  {
-    id: uuidv4(),
-    name: "MANJARI BRINJAL",
-    image: "https://farmagritech.com/wp-content/uploads/2020/03/MANJARI-.jpg",
-    price: 140.00,
-    material: "seed",
-    brand: "Manjari",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: 4,
-    offer: 10,
-    idealFor:"brinjal" ,
-    color: "purple"
-  },
-  {
-    id: uuidv4(),
-    name: "MEGHANA BRINJAL",
-    image: "https://farmagritech.com/wp-content/uploads/2019/12/Photo-from-siddaraddi-ramaraddi-20.jpg",
-    price:140.00,
-    material: "seed",
-    brand: "Meghana",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: 3.8,
-    offer: 10,
-    idealFor:"brinjal" ,
-    color: "purple"
-  },
-
-  /*----------------------cauliflower-seed---------------*/
-
-  {
-    id: uuidv4(),
-    name: "CFL REGAL WHITE CAULIFLOWER",
-    image: "https://farmagritech.com/wp-content/uploads/2020/05/CFL-REGAL-WHITE-CAULIFLOWER.png",
-    price: 180.00,
-    material: "seed",
-    brand: "CFL Regal",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: 4,
-    offer: 10,
-    idealFor:"cauliflower" ,
-    color: "white"
-  },
-  {
-    id: uuidv4(),
-    name: "SUHASINI CAULIFLOWER SEEDS",
-    image: "https://farmagritech.com/wp-content/uploads/2020/05/SUHASINI-CAULIFLOWER-SYNGENTA.png",
-    price: 185.00,
-    material: "seed",
-    brand: "Suhasini",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: 4.2,
-    offer: 6,
-    idealFor:"cauliflower" ,
-    color: "white"
-  },
-  {
-    id: uuidv4(),
-    name: "TETRIS CAULIFLOWER SEEDS",
-    image: "https://farmagritech.com/wp-content/uploads/2020/05/TETRIS-CAULIFLOWER.png",
-    price: 183.00,
-    material: "seed",
-    brand: "Tetris",
-    inStock: true,
-    isnew: true,
-    fastDelivery: true,
-    ratings: 4.5,
-    offer: 6,
-    idealFor:"cauliflower" ,
-    color: "white"
-  },
-/*--------------------chilli-seed------------*/
-{
-  id: uuidv4(),
-  name: "Chilli NS 1701",
-  image: "https://farmagritech.com/wp-content/uploads/2020/04/Chilli-1701.png",
-  price: 173.00,
-  material: "seed",
-  brand: "Namdhari",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4.1,
-  offer: 6,
-  idealFor:"chilli" ,
-  color: "red"
-},
-
-/*-----------------corn-seed---------*/
-{
-  id: uuidv4(),
-  name: "DHANYA MAIZE SEEDS MM9222",
-  image: "https://farmagritech.com/wp-content/uploads/2020/07/20200612_174746-1-rotated-e1593949519303.jpg",
-  price: 1870.00,
-  material: "seed",
-  brand: "Dhanya",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.7,
-  offer: 15,
-  idealFor:"corn" ,
-  color: "yellow"
-},
-/*-------------------------marigold-seed------------*/
-{
-  id: uuidv4(),
-  name: "ARADHANA MARIGOLD",
-  image: "https://farmagritech.com/wp-content/uploads/2020/03/indus-aradhana.png",
-  price: 2670.00,
-  material: "seed",
-  brand: "Aradhana",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4.7,
-  offer: 20,
-  idealFor:"marigold" ,
-  color: "yellow"
-},
-{
-  id: uuidv4(),
-  name: "ARROW GOLD MARIGOLD",
-  image: "https://farmagritech.com/wp-content/uploads/2020/03/arrow-gold-f1.png",
-  price: 1660.00,
-  material: "seed",
-  brand: "Arrow gold",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4.3,
-  offer: 15,
-  idealFor:"marigold" ,
-  color: "yellow"
-},
- 
-/*--------------------------watermelon-----------*/
-{
-  id: uuidv4(),
-  name: "JUMBO QUEEN WATER MELON",
-  image: "https://farmagritech.com/wp-content/uploads/2020/03/water-melon-Jumbo-queen.png",
-  price: 1600.00,
-  material: "seed",
-  brand: "Jumbo queen",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4,
-  offer: 15,
-  idealFor:"watermelon" ,
-  color: "red"
-},
-{
-  id: uuidv4(),
-  name: "KIRAN 2 WATERMELON",
-  image: "https://farmagritech.com/wp-content/uploads/2020/03/KIRAN-2.jpg",
-  price: 1620.00,
-  material: "seed",
-  brand: "Kiran",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.5,
-  offer: 10,
-  idealFor:"watermelon" ,
-  color: "red"
-},
-/*-----------------tomatos----------------------------*/
-{
-  id: uuidv4(),
-  name: "ABHINAV TOMATO",
-  image: "https://farmagritech.com/wp-content/uploads/2019/11/abhinav.png",
-  price: 985.00,
-  material: "seed",
-  brand: "Syngenta",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.5,
-  offer: 10,
-  idealFor:"tomato" ,
-  color: "red"
-},
-{
-  id: uuidv4(),
-  name: "HARD ROCK (TO-8011) TOMATO",
-  image: "https://farmagritech.com/wp-content/uploads/2019/11/HARD-ROCK.jpg",
-  price: 885.00,
-  material: "seed",
-  brand: "Syngenta",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.9,
-  offer: 20,
-  idealFor:"tomato" ,
-  color: "red"
-},
-/*-------------------------------insectisides------------*/
-
-{
-  id: uuidv4(),
-  name: "ADMIRE INSECTICIDE",
-  image: "https://farmagritech.com/wp-content/uploads/2020/03/I-Admire.jpg",
-  price: 315.00,
-  material: "pesticide",
-  brand: "Bayer",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.9,
-  offer: 15,
-  idealFor:"insect" ,
-  color: "white"
-},
-{
-  id: uuidv4(),
-  name: "ANTRACOL",
-  image: "https://krushikendra.com/image/cache/catalog/Bayer/antracol-800x800.png",
-  price: 225.00,
-  material: "pesticide",
-  brand: "Bayer",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 3.9,
-  offer: 15,
-  idealFor:"insect" ,
-  color: "white"
-},
-
-/* --------------------------fertilizer-------------------*/
-
-{
-  id: uuidv4(),
-  name: "Agromin - 500 ml ",
-  image: "https://farmkey.in/uploads/product/Agromin.png",
-  price: 232.00,
-  material: "fertilizer",
-  brand: "Aries agro limited",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4,
-  offer: 10,
-  idealFor:"micronutrients" ,
-  color: "white"
-},
-/*------------------accesories------------------------*/
-{
-  id: uuidv4(),
-  name: "3 in 1 Battery Sprayer (12x12) ",
-  image: "https://krushikendra.com/image/cache/catalog/sprayers%20offer%20photos/3%20in%201%20Battery12x8-800x800.jpg",
-  price: 4200,
-  material: "accesories",
-  brand: "National Pesticides and Chemicals",
-  inStock: true,
-  isnew: true,
-  fastDelivery: true,
-  ratings: 4,
-  offer: 20,
-  idealFor:"spray" ,
-  color: "orange"
-},
-
-]
-
