+++ frontend/src/index.js
@@ -1,9 +1,9 @@
 import React, { StrictMode } from "react";
 import ReactDOM from "react-dom";
-import { ReducerProvider } from "./Reducer-context";
+import { ReducerProvider } from "./Reducer-context/Reducer-context";
 import { BrowserRouter as Router } from "react-router-dom";
 import App from "./App";
-import { LoginProvider } from "./LoginContext";
+import { LoginProvider } from "./Reducer-context/LoginContext";
 const rootElement = document.getElementById("root");
 ReactDOM.render(
   <StrictMode>
