+++ frontend/src/pages/Accessories.js
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
-import { useReduce } from "./Reducer-context";
+import { useReduce } from "../Reducer-context/Reducer-context";
 import { Link } from "react-router-dom";
-import Product from "./Products";
-export default function Men({
+import { Products } from "./Products";
+export function Accessories({
   filteredData,
   Add_to_cart_button,
   Add_to_wishlist_button
@@ -11,11 +11,11 @@ export default function Men({
   useEffect(() => {
     dispatch({ type: "OFFER", payload: 0 });
     dispatch({ type: "NEW", payload: false });
-    dispatch({ type: "MATERIAL", payload: "seed" });
+    dispatch({ type: "MATERIAL", payload: "accesories" });
   }, []);
   return (
     <>
-      <Product
+      <Products
         filteredData={filteredData}
         Add_to_cart_button={Add_to_cart_button}
         Add_to_wishlist_button={Add_to_wishlist_button}
