+++ frontend/src/pages/Signup.js
@@ -1,31 +1,48 @@
-import { useLogin } from "./LoginContext";
+import { useLogin } from "../Reducer-context/LoginContext";
 import { useNavigate, useLocation, Link } from "react-router-dom";
 import { useEffect, useState } from "react";
-import treebanner from "./assets/treebanner.png";
-import { useReduce } from "./Reducer-context";
-export default function Login() {
-  let { isUserLogIn, LoginWithCredentials } = useLogin();
-  const { dispatch, user } = useReduce();
+import axios from "axios";
+import treebanner from "../assets/treebanner.png";
+import { useReduce } from "../Reducer-context/Reducer-context";
+
+export function Signup() {
+  let { isUserLogIn } = useLogin();
+  let { dispatch } = useReduce();
+  let [error, setError] = useState("");
+  const [userName, setUserName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   let navigate = useNavigate();
   let { state } = useLocation();
-  let [error, setError] = useState("");
+  console.log("form", state);
+
   useEffect(() => {
     if (isUserLogIn) {
       navigate(state?.from ? state.from : "/", { replace: true });
     }
   }, [isUserLogIn]);
 
-  async function LoginHandler() {
-    dispatch({ type: "LOAD", payload: true });
-    let errorpassed = await LoginWithCredentials(email, password);
-    setError(errorpassed);
-    dispatch({ type: "LOAD", payload: false });
+  async function signupHandler() {
+    try {
+      dispatch({ type: "LOAD", payload: true });
+      let response = await axios.post(
+        "https://ecomm-demo-1.utpalpati.repl.co/signup",
+        { user: { userName, email, password } }
+      );
+      if (response.status === 200) {
+        setError("");
+        navigate("/login");
+      }
+      dispatch({ type: "LOAD", payload: false });
+    } catch (error) {
+      setError(error.response.data.message);
+      dispatch({ type: "LOAD", payload: false });
+      console.log("signuphandler error");
+    }
   }
   return (
-    <div className="login">
-      <div className="login-sideimg">
+    <div className="signup">
+      <div className="signup-sideimg">
         {" "}
         <img
           src={treebanner}
@@ -35,15 +52,26 @@ export default function Login() {
       </div>
 
       <div className="form-container">
-        <h3>Login</h3>
+        <h3>Signup</h3>
 
         <form
           onSubmit={(e) => {
             e.preventDefault();
-            LoginHandler();
+            signupHandler();
           }}
           className="form"
         >
+          <label class="input md">
+            <input
+              type="text"
+              onChange={(e) => {
+                setUserName(e.target.value);
+              }}
+              class="input-text"
+              required
+            />
+            <span class="placeholder">Name</span>
+          </label>
           <label class="input md">
             <input
               type="email"
@@ -64,14 +92,14 @@ export default function Login() {
               class="input-text"
               required
             />
-            <span class="placeholder">Password</span>
+            <span class="placeholder">password</span>
           </label>
           <div className="form-action">
             <button class="secondary-button md">
-              <Link to="/signup">Signup</Link>
+              <Link to="/login">Login</Link>
             </button>
             <button type="submit" class="secondary-button md">
-              Login
+              Signup
             </button>
           </div>
         </form>
