+++ frontend/src/pages/Tenpercent.js
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
-import { useReduce } from "./Reducer-context";
+import { useReduce } from "../Reducer-context/Reducer-context";
 import { Link } from "react-router-dom";
-import Product from "./Products";
-export default function Fifteenpercent({
+import { Products } from "./Products";
+export function Tenpercent({
   filteredData,
   Add_to_cart_button,
   Add_to_wishlist_button
@@ -11,11 +11,11 @@ export default function Fifteenpercent({
   useEffect(() => {
     dispatch({ type: "MATERIAL", payload: "All" });
     dispatch({ type: "NEW", payload: false });
-    dispatch({ type: "OFFER", payload: 15 });
+    dispatch({ type: "OFFER", payload: 10 });
   }, []);
   return (
     <>
-      <Product
+      <Products
         filteredData={filteredData}
         Add_to_cart_button={Add_to_cart_button}
         Add_to_wishlist_button={Add_to_wishlist_button}
