+++ frontend/src/pages/index.js
@@ -1,0 +1,18 @@
+export { Home } from "./Home";
+export { Accessories } from "./Accessories";
+export { Cart } from "./Cart";
+export { Fertilizer } from "./Fertilizer";
+export { Login } from "./Login";
+export { Signup } from "./Signup";
+export { New } from "./New";
+export { Pesticide } from "./Pesticide";
+export { ProductPage } from "./ProductPage";
+export { Products } from "./Products";
+export { Seed } from "./Seed";
+export { Sixpercent } from "./Sixpercent";
+export { Tenpercent } from "./Tenpercent";
+export { Twentypercent } from "./Twentypercent";
+export { Fifteenpercent } from "./Fifteenpercent";
+export { Wishlist } from "./Wishlist";
+export { User } from "./User";
+export { NotFound } from "./NotFound";
